+++ packages/build-tool/config/webpack.config.ts
@@ -1,16 +1,26 @@
 import path from 'path';
 import { cwd } from 'node:process';
-import webpack from 'webpack';
+import webpack, { type PathData } from 'webpack';
+import { Configuration as WebpackDevServerConfiguration } from 'webpack-dev-server';
 
 import CopyWebpackPlugin from 'copy-webpack-plugin';
 import MiniCssExtractPlugin from 'mini-css-extract-plugin';
 import { CleanWebpackPlugin } from 'clean-webpack-plugin';
 
 // eslint-disable-next-line import/no-unresolved
-import { getEntries, processFilename, type PathData } from '../utils/webpack';
+import { getEntries, processFilename } from '../utils/webpack';
+
+interface WPScriptsConfig extends webpack.Configuration {
+  devServer?: WebpackDevServerConfiguration;
+  resolve: {
+    alias: {
+      [x: string]: string;
+    };
+  };
+}
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const defaultConfig: webpack.Configuration = require('@wordpress/scripts/config/webpack.config');
+const defaultConfig: WPScriptsConfig = require('@wordpress/scripts/config/webpack.config');
 
 /**
  * Check if the build is running in production mode.
@@ -44,12 +54,14 @@ const mode: string = isProduction ? 'production' : 'development';
  * @see https://github.com/WordPress/gutenberg/tree/trunk/packages/scripts#extending-the-webpack-config
  * @see https://github.com/WordPress/gutenberg/blob/trunk/packages/scripts/config/webpack.config.js
  */
-const config = (): webpack.Configuration => ({
+const config = (): webpack.Configuration => (defaultConfig ? {
   ...defaultConfig,
 
   // Dynamically produce entries from the slotfills index file and all blocks.
   entry: () => {
-    const blocks = defaultConfig.entry();
+    let blocks = typeof defaultConfig.entry === 'function' ? defaultConfig.entry() : {};
+    blocks = blocks && typeof blocks === 'object' ? blocks : {};
+
     const entries = blocksOnly === true ? {} : getEntries(entriesDir);
 
     return {
@@ -76,7 +88,7 @@ const config = (): webpack.Configuration => ({
 
   // Configure plugins.
   plugins: [
-    ...defaultConfig.plugins,
+    ...(Array.isArray(defaultConfig.plugins) ? defaultConfig.plugins : []),
     new CopyWebpackPlugin({
       patterns: [
         {
@@ -127,6 +139,6 @@ const config = (): webpack.Configuration => ({
       directory: '/build',
     },
   },
-});
+} : {});
 
 export default config;