+++ packages/build-tool/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.1.3 -- @alleyinteractive/build-tool
+
+- In the `tsconfig.esm.json` file, when `moduleResolution` is set to `NodeNext`, the `module` field must be set to `NodeNext` as well.
+- Updated webpack and addressed TypeScript errors.
+
 ## v0.1.2 -- @alleyinteractive/build-tool
 
 - Add exports for webpack config and update README with instructions on how to extend the config.